feat(POST): POST endopoint now works and it included validations and error handling

+++ src/main/java/com/pikolinc/meliecommerce/domain/dto/item/ItemResponseDTO.java
@@ -4,6 +4,6 @@ public record ItemResponseDTO(
         Long id,
         String name,
         String description,
-        Float price
+        Double price
 ) {
 }
